+++ test/fakeString.h
@@ -23,7 +23,7 @@ public:
 	String & operator + (unsigned int num)		{return (*this);}
 	String & operator + (long num)			{return (*this);}
 	String & operator + (unsigned long num)	{return (*this);}
-	String & operator + (float num)		{return (*this);}
+	String & operator + (double num)		{return (*this);}
 	String & operator + (double num)		{return (*this);}
 
     const char* c_str() const;
@@ -32,7 +32,7 @@ public:
 	int indexOf( char ch, unsigned int fromIndex ) const;
     String substring( unsigned int beginIndex ) const { return substring(beginIndex, this->length); };
 	String substring( unsigned int beginIndex, unsigned int endIndex ) const;
-    float toFloat() const;
+    double toFloat() const;
     int toInt() const;
 };
 #endif
